pyproject.toml
[project]
name = "ticketing"
version = "0.1.0"
description = "Django-based ticketing system similar to Jira"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.0,<5.3",
    "django-crispy-forms>=2.1",
    "crispy-bootstrap5>=2023.10",
    "djangorestframework>=3.15.0",
    "pillow>=10.1.0",
    "psycopg>=3.1.12,<4.0", # PostgreSQL adapter (optional)
    "psycopg-pool>=3.2.0", # Connection pooling for PostgreSQL (optional)
    "mysqlclient>=2.2.1", # MySQL adapter (optional)
    "drf-yasg>=1.21.7", # API documentation (optional)
    "whitenoise>=6.6.0", # Static file serving
    "gunicorn>=21.2.0", # WSGI HTTP Server (for production)
    "python-dotenv>=1.0.0", # Environment variable management
    "psycopg2-binary",
    "dj-database-url",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "ARG", "SIM"]
ignore = ["E203", "E501"]
